Fixes frequency calculation for non 32 bit platforms.
With Arduino, setFrequency(400 * 1000 * 1000) was calculated as a 16 bit
signed integer. This resulted in a wrongly set frequency. This fixes the
calculation by ensuring a 32 unsigned integer is used.

+++ examples/BusyMan/BusyMan.ino
@@ -110,7 +110,7 @@ void setup(){
 
     rfm.setBufferSize(10);   // set the internal buffer size.
     rfm.setPacketLength(4); // set the packet length.
-    rfm.setFrequency(434*1000*1000); // set the frequency.
+    rfm.setFrequency((uint32_t) 434*1000*1000); // set the frequency.
 
     // baudrate is default, 4800 bps now.
