+++ examples/Maximum_PingPong/Maximum_PingPong.ino
@@ -191,7 +191,7 @@ void setup(){
 
     rfm.setBufferSize(3);   // set the internal buffer size.
     rfm.setPacketLength(64); // set the packet length.
-    rfm.setFrequency(434*1000*1000); // set the frequency.
+    rfm.setFrequency((uint32_t) 434*1000*1000); // set the frequency.
 
 
     // rfm.baud9600();
