+++ examples/PingPong/PingPong.ino
@@ -153,7 +153,7 @@ void setup(){
 
     rfm.setBufferSize(2);   // set the internal buffer size.
     rfm.setPacketLength(6); // set the packet length.
-    rfm.setFrequency(434*1000*1000); // set the frequency.
+    rfm.setFrequency((uint32_t) 434*1000*1000); // set the frequency.
 
     // by default use 4800 bps
